requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.159
botocore==1.12.159
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.1.1
colorama==0.4.1
constantly==15.1.0
crochet==1.9.0
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
DateTime==4.3
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.7
django-background-tasks==1.2.0
django-compat==1.0.15
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.7.1
docutils==0.14
entrypoints==0.3
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.17
incremental==17.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
keyring==19.0.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.3
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
numpy==1.16.4
numpydoc==0.9.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.5.1
parso==0.4.0
path.py==12.0.1
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psutil==5.6.2
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
PyQt5==5.12
PyQt5-sip==4.19.17
PyQtWebEngine==5.12.1
pyrsistent==0.15.2
python-dateutil==2.8.0
python-dotenv==0.10.2
python-scrapyd-api==2.1.2
pytz==2019.1
                                
pyzmq==18.0.1
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.7.1
queuelib==1.5.0
requests==2.22.0
rope==0.14.0
s3transfer==0.2.0
scipy==1.3.0
Scrapy==1.6.0
scrapy-djangoitem==1.1.1
scrapyd==1.2.0
seaborn==0.9.0
service-identity==18.1.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.3
spyder-kernels==0.4.2
sqlparse==0.3.0
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
Twisted==19.2.0
typed-ast==1.3.5
urllib3==1.25.3
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==7.0
whitenoise==4.1.2
wrapt==1.11.1
zipp==0.5.1
zope.interface==4.6.0
